ignore `p2::sync::udp_send_to_closed_receiver` test on Windows
Alex and I spent some time trying to debug this before concluding that it wasn't
worth spending more time on.  The circumstances under which it fails are both
narrow and mystifying, and we don't want to block overall improved test coverage
on that case.

Specifically, it only fails when all three variations (async, sync + thread
blocking, and sync + no thread blocking) are run at the same time, and only on
Windows.  Running any two of those variations on Windows works fine.  :shrug:
